Added datastore tests and test remote user database.

+++ tests/test_isolatedb.py
@@ -33,11 +33,13 @@ from bigsdb.constants import CONNECTION_DETAILS
 from bigsdb.datastore import Datastore
 from bigsdb.xml_parser import XMLParser
 
-TEST_DATABASE = "bigsdb_test_isolates2"
+TEST_ISOLATE_DATABASE = "bigsdb_test_isolates"
+TEST_USERS_DATABASE = "bigsdb_test_users"
 HOST = "localhost"
 PORT = 5432
 USER = "bigsdb_tests"
 PASSWORD = "test"
+PERSIST = True  # Set to False to drop and recreate test database each time.
 
 
 class TestIsolateDB(unittest.TestCase):
@@ -49,33 +51,107 @@ class TestIsolateDB(unittest.TestCase):
         self.assertEqual(len(eav_fieldnames), 4)
         self.assertTrue("Bexsero_reactivity" in eav_fieldnames)
 
+    def test_get_eav_fields(self):
+        fields = self.datastore.get_eav_fields()
+        self.assertTrue(len(fields) > 1)
+        all_have_name = True
+        all_have_type = True
+        for field in fields:
+            if not field.get("field"):
+                all_have_name = False
+            if not field.get("value_format"):
+                all_have_type = False
+        self.assertTrue(all_have_name and all_have_type)
+
+    def test_isolates_with_seqbin(self):
+        ids, labels = self.datastore.get_isolates_with_seqbin()
+        self.assertEqual(len(ids), 3)
+        self.assertEqual(labels.get(1, ""), "1) A4/M1027")
+
+    def test_get_remote_user_info(self):
+        user_info = self.datastore.get_remote_user_info("jdoe", 1)
+        self.assertEqual(user_info.get("first_name"), "John")
+        self.assertEqual(user_info.get("surname"), "Doe")
+        self.assertEqual(user_info.get("email"), "john.doe@test.com")
+
+    def test_get_seqbin_count(self):
+        count = self.datastore.get_seqbin_count()
+        self.assertEqual(count, 3)
+
+    def test_get_user_info_from_username(self):
+        user_info = self.datastore.get_user_info_from_username("test")
+        self.assertEqual(user_info.get("first_name"), "Test")
+        self.assertEqual(user_info.get("surname"), "User")
+        self.assertEqual(user_info.get("email"), "test@test.com")
+        remote_user_info = self.datastore.get_user_info_from_username("jdoe")
+        self.assertEqual(remote_user_info.get("email"), "john.doe@test.com")
+
+    def test_isolate_exists(self):
+        self.assertTrue(self.datastore.isolate_exists(1))
+        self.assertFalse(self.datastore.isolate_exists(1200))
+
+    def test_run_query(self):
+        list = self.datastore.run_query(
+            "SELECT id FROM isolates", None, {"fetch": "col_arrayref"}
+        )
+        self.assertTrue(len(list) > 1)
+        empty_list = self.datastore.run_query(
+            "SELECT id FROM isolates WHERE id > 1000", None, {"fetch": "col_arrayref"}
+        )
+        self.assertTrue(len(empty_list) == 0)
+        # TODO row_arrayref
+        # TODO row_hashref
+        # TODO all_hashref
+        # TODO all_arrayref
+
     @classmethod
     def setUpClass(cls):
-        # Setup test isolate database
-
         cls.con = psycopg2.connect(dbname="postgres")
-        cls.con.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
-        cur = cls.con.cursor()
-        cur.execute(f"DROP DATABASE IF EXISTS {TEST_DATABASE}")
-        cur.execute(f"DROP USER IF EXISTS {USER}")
-        cur.execute(f"CREATE USER {USER}")
-        cur.execute(f"ALTER USER {USER} WITH PASSWORD '{PASSWORD}'")
-
-        cur.execute(f"CREATE DATABASE {TEST_DATABASE}")
-        cls.con.commit()
-        cls.con.close()
-        cls.con = psycopg2.connect(dbname=TEST_DATABASE)
-        cur = cls.con.cursor()
         dir = pathlib.Path(__file__).parent.resolve()
-        with open(f"{dir}/databases/bigsdb_test_isolates.sql", "r") as f:
-            cur.copy_expert(sql=f.read(), file=f)
-        cur.execute(
-            "GRANT SELECT, INSERT, UPDATE, DELETE ON ALL TABLES "
-            f"IN SCHEMA public TO {USER}"
-        )
-        cls.con.commit()
-        cur.close()
-        cls.con.close()
+        drop_and_recreate = True
+        if PERSIST:
+            isolatedb_exists = database_exists(cls.con, TEST_ISOLATE_DATABASE)
+            userdb_exists = database_exists(cls.con, TEST_USERS_DATABASE)
+            if isolatedb_exists and userdb_exists:
+                drop_and_recreate = False
+
+        if drop_and_recreate:  # Setup test isolate database
+            cls.con.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
+            cur = cls.con.cursor()
+            cur.execute(f"DROP DATABASE IF EXISTS {TEST_ISOLATE_DATABASE}")
+            cur.execute(f"DROP DATABASE IF EXISTS {TEST_USERS_DATABASE}")
+            cur.execute(f"DROP USER IF EXISTS {USER}")
+            cur.execute(f"CREATE USER {USER}")
+            cur.execute(f"ALTER USER {USER} WITH PASSWORD '{PASSWORD}'")
+
+            cur.execute(f"CREATE DATABASE {TEST_ISOLATE_DATABASE}")
+            cur.execute(f"CREATE DATABASE {TEST_USERS_DATABASE}")
+            cls.con.commit()
+            cls.con.close()
+            cls.con = psycopg2.connect(dbname=TEST_ISOLATE_DATABASE)
+            cur = cls.con.cursor()
+
+            with open(f"{dir}/databases/bigsdb_test_isolates.sql", "r") as f:
+                cur.copy_expert(sql=f.read(), file=f)
+            cur.execute(
+                "GRANT SELECT, INSERT, UPDATE, DELETE ON ALL TABLES "
+                f"IN SCHEMA public TO {USER}"
+            )
+            cls.con.commit()
+            cur.close()
+            cls.con.close()
+            cls.con = psycopg2.connect(dbname=TEST_USERS_DATABASE)
+            cur = cls.con.cursor()
+
+            with open(f"{dir}/databases/bigsdb_test_users.sql", "r") as f:
+                cur.copy_expert(sql=f.read(), file=f)
+            cur.execute(
+                "GRANT SELECT, INSERT, UPDATE, DELETE ON ALL TABLES "
+                f"IN SCHEMA public TO {USER}"
+            )
+            cls.con.commit()
+            cur.close()
+            cls.con.close()
 
         # Read BIGSdb config file
         conf_file = f"{dir}/config_files/bigsdb.conf"
@@ -94,13 +170,19 @@ class TestIsolateDB(unittest.TestCase):
         # Connect
         cls.data_connector = DataConnector(system=cls.system, config=cls.config)
         cls.db = cls.data_connector.get_connection(
-            dbase_name=TEST_DATABASE,
+            dbase_name=TEST_ISOLATE_DATABASE,
             host=HOST,
             port=PORT,
             user=USER,
             password=PASSWORD,
         )
-
+        cls.user_db = cls.data_connector.get_connection(
+            dbase_name=TEST_USERS_DATABASE,
+            host=HOST,
+            port=PORT,
+            user=USER,
+            password=PASSWORD,
+        )
         # Set up datastore
         cls.datastore = Datastore(
             db=cls.db,
@@ -108,12 +190,25 @@ class TestIsolateDB(unittest.TestCase):
             config=cls.config,
             parser=cls.parser,
         )
+        cls.datastore.add_user_db(1, cls.user_db, TEST_USERS_DATABASE)
 
     @classmethod
     def tearDownClass(cls):
         cls.db.close()
+        cls.user_db.close()
+        if PERSIST:
+            return
         cls.con = psycopg2.connect(dbname="postgres")
         cls.con.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
         cur = cls.con.cursor()
-        cur.execute(f"DROP DATABASE {TEST_DATABASE}")
+        cur.execute(f"DROP DATABASE {TEST_ISOLATE_DATABASE}")
         cur.execute(f"DROP USER {USER}")
+
+
+def database_exists(conn, db_name):
+    cursor = conn.cursor()
+    query = "SELECT 1 FROM pg_database WHERE datname = %s"
+    cursor.execute(query, (db_name,))
+    result = cursor.fetchone()
+    cursor.close()
+    return result is not None
